% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reemsplots2-package.R
\docType{package}
\name{reemsplots2}
\alias{reemsplots2}
\alias{reemsplots2-package}
\title{reemsplots2}
\description{
This package implements several functions to inspect and visualize the results of analyzing geo-referenced genetic data with EEMS as well as to evaluate the EEMS model fit.
\describe{
\item{make_eems_plots}{Produce contour plots of the effective migration and diversity
rates, as well as scatter plots of observed vs fitted genetic dissimilarities.}
\item{plot_population_grid}{Plot the population grid constructed by EEMS.}
\item{plot_voronoi_tiles}{Takes random samples from the posterior distribution of the migration and diversity rates to give an idea of the posterior variance.}
\item{plot_resid_heatmap}{Computes the matrix of residuals (differences between the observed and fitted dissimilarities) and plots it as a heat map.}
}
In addition to the help pages, see the README page on \href{https://github.com/dipetkov/eems}{github} for examples.
}
